move relay command class to separate file
