+++ src/core/Microsoft.PowerToys.Settings.UI.Lib/ViewModels/Commands/RelayCommand.cs
@@ -31,31 +31,4 @@ namespace Microsoft.PowerToys.Settings.UI.Lib.ViewModels.Commands
 
         public void OnCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
     }
-
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:File may only contain a single type", Justification = "abstract T and abstract")]
-    public class RelayCommand<T> : ICommand
-    {
-        private readonly Action<T> execute;
-
-        private readonly Func<T, bool> canExecute;
-
-        public event EventHandler CanExecuteChanged;
-
-        public RelayCommand(Action<T> execute)
-            : this(execute, null)
-        {
-        }
-
-        public RelayCommand(Action<T> execute, Func<T, bool> canExecute)
-        {
-            this.execute = execute ?? throw new ArgumentNullException(nameof(execute));
-            this.canExecute = canExecute;
-        }
-
-        public bool CanExecute(object parameter) => canExecute == null || canExecute((T)parameter);
-
-        public void Execute(object parameter) => execute((T)parameter);
-
-        public void OnCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
-    }
 }
